test(common-ts): share one divergence map between the corpora
Both characterization suites carried the same runner verbatim.
